feat(spec): add categorical inputs and record lookups
Add reusable named input definitions with optional enum types, so a function
can bind a categorical input by name while retaining its own symbol, unit,
domain, and description. Keep enum definitions focused on their stable members
and canonical values rather than duplicating binding-specific metadata.

Add typed record lookups keyed by enum inputs. Validate that a lookup references
an in-scope categorical input, covers every enum member exactly once, and
provides exactly the fields required by its declared record output. Compile
lookups into the semantic IR as explicit conversions from categorical values to
numeric record fields.

Lower categorical inputs and lookups consistently across Rust, C, C++, and
Python. Generate target-native enum types and lookup expressions, namespace C
enum identifiers by enum type, and retain the existing public APIs and
documentation contracts. Extend the NumPy extension and Python wrappers with
uint32-backed enum arrays, broadcasting, scalar dispatch, typing, generated
enum documentation, and tests.

+++ targets/ptfkit-py/src/ptfkit/_dispatch.py
@@ -2,11 +2,11 @@ from __future__ import annotations
 
 from typing import Any
 
-from numpy import asarray, float64, ufunc
+from numpy import asarray, ufunc
 
 
 def call(function: ufunc, *inputs: object, out: object) -> Any:  # noqa: ANN401
-    inputs = tuple(asarray(value, dtype=float64) for value in inputs)
+    inputs = tuple(asarray(value) for value in inputs)
     if out is None:
         return function(*inputs)
     if isinstance(out, tuple):
